Updated ADAMS to allow non-int Z (wf class still uses int)

+++ src/ADAMS_solveLocalBS.cpp
@@ -22,7 +22,7 @@ bool debug=false; //if true, will print progress messages.
 
 //******************************************************************************
 int solveDBS(std::vector<double> &p, std::vector<double> &q, double &en,
-    std::vector<double> v, int Z, int n, int ka,
+    std::vector<double> v, double Z, int n, int ka,
     std::vector<double> r, std::vector<double> drdt, double h, int NGP,
     int &pinf, int &its, double &eps, double alpha, int log_dele_or)
 /*
@@ -307,7 +307,7 @@ an error.
 
 //******************************************************************************
 int outwardAM(std::vector<double> &p, std::vector<double> &q, double &en,
-    std::vector<double> v, int Z, int ka,
+    std::vector<double> v, double Z, int ka,
     std::vector<double> r, std::vector<double> drdt, double h, int NGP,
     int nf, double alpha)
 /*
